Revert "1. OOP 2. MimeType Link Icons"
