--- README-TEMPLATE.md
@@ -1,30 +1,0 @@
-# Name des Plugins
-WordPress-Plugin: Kurzbeschreibung des Plugins.
-
-
-## Download
-
-GITHub-Repo: https://github.com/RRZE-Webteam/rrze-...
-
-## Autor
-RRZE-Webteam , http://www.rrze.fau.de
-
-## Copryright
-GNU General Public License (GPL) Version 2
-
-
-## Beschreibung
-
-Ausführliche Beschreibung des Plugins.
-
-## Einstellungen
-
-Beschreibung der einzelnen Einstellungen des Plugins.
-
-## Verwendung
-
-Hinweise zur Verwendung der einzelnen Funktionen des Plugins (Shortcodes, Widgets, etc.) inklusive Anwendungsbeispiele.
-
-```html
-[mein_shortcode]
-```
