--- assets/css/basis-shortcode.css
@@ -1,4 +1,0 @@
-@charset "UTF-8";
-/**
-Shortcode Style
-*/
