--- includes/Exception.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace RRZE\Downloads;
-
-defined('ABSPATH') || exit;
-
-/**
- * Exception-Klasse
- */
-class Exception extends \Exception
-{
-}
