+++ languages/rrze-downloads-en_US.po
@@ -42,95 +42,95 @@ msgstr ""
 "Your Wordpress version is %s. Please upgrade at least to Wordpress version "
 "%s."
 
-#: ../includes/Shortcode.php:219
+#: ../shortcodes/downloads.php:219
 msgid "Dowloads: Es konnten keine Dateien gefunden werden."
 msgstr "Downloads: No files could be found."
 
-#: ../assets/taxonomies/attachment-category.php:10
-#: ../assets/taxonomies/attachment-category.php:20
+#: ../taxonomies/attachment-category.php:10
+#: ../taxonomies/attachment-category.php:20
 msgid "Kategorien"
 msgstr "Categories"
 
-#: ../assets/taxonomies/attachment-category.php:11
+#: ../taxonomies/attachment-category.php:11
 msgid "Kategorie"
 msgstr "Category"
 
-#: ../assets/taxonomies/attachment-category.php:12
+#: ../taxonomies/attachment-category.php:12
 msgid "Kategorien suchen"
 msgstr "Search Categories"
 
-#: ../assets/taxonomies/attachment-category.php:13
-#: ../assets/taxonomies/attachment-category.php:49
-#: ../assets/taxonomies/attachment-category.php:92
+#: ../taxonomies/attachment-category.php:13
+#: ../taxonomies/attachment-category.php:49
+#: ../taxonomies/attachment-category.php:92
 msgid "Alle Kategorien"
 msgstr "All Categories"
 
-#: ../assets/taxonomies/attachment-category.php:14
+#: ../taxonomies/attachment-category.php:14
 msgid "Übergeordnet"
 msgstr "Parent"
 
-#: ../assets/taxonomies/attachment-category.php:15
+#: ../taxonomies/attachment-category.php:15
 msgid "Übergeordnet:"
 msgstr "Parent:"
 
-#: ../assets/taxonomies/attachment-category.php:16
+#: ../taxonomies/attachment-category.php:16
 msgid "Kategorie bearbeiten"
 msgstr "Edit Category"
 
-#: ../assets/taxonomies/attachment-category.php:17
+#: ../taxonomies/attachment-category.php:17
 msgid "Kategorie aktualisieren"
 msgstr "Update"
 
-#: ../assets/taxonomies/attachment-category.php:18
+#: ../taxonomies/attachment-category.php:18
 msgid "Neue Kategorie hinzufügen"
 msgstr "Add New Category"
 
-#: ../assets/taxonomies/attachment-category.php:19 ../assets/taxonomies/attachment-tag.php:20
+#: ../taxonomies/attachment-category.php:19 ../taxonomies/attachment-tag.php:20
 msgid "Name"
 msgstr "Name"
 
-#: ../assets/taxonomies/attachment-tag.php:10 ../assets/taxonomies/attachment-tag.php:24
+#: ../taxonomies/attachment-tag.php:10 ../taxonomies/attachment-tag.php:24
 msgid "Schlagwörter"
 msgstr "Tags"
 
-#: ../assets/taxonomies/attachment-tag.php:11
+#: ../taxonomies/attachment-tag.php:11
 msgid "Schlagwort"
 msgstr "Tag"
 
-#: ../assets/taxonomies/attachment-tag.php:12
+#: ../taxonomies/attachment-tag.php:12
 msgid "Schlagwörter suchen"
 msgstr "Search Tags"
 
-#: ../assets/taxonomies/attachment-tag.php:13
+#: ../taxonomies/attachment-tag.php:13
 msgid "Beliebte Schlagwörter"
 msgstr "Most Used"
 
-#: ../assets/taxonomies/attachment-tag.php:14 ../assets/taxonomies/attachment-tag.php:53
-#: ../assets/taxonomies/attachment-tag.php:96
+#: ../taxonomies/attachment-tag.php:14 ../taxonomies/attachment-tag.php:53
+#: ../taxonomies/attachment-tag.php:96
 msgid "Alle Schlagwörter"
 msgstr "All Tags"
 
-#: ../assets/taxonomies/attachment-tag.php:17
+#: ../taxonomies/attachment-tag.php:17
 msgid "Schlagwort bearbeiten"
 msgstr "Edit Tag"
 
-#: ../assets/taxonomies/attachment-tag.php:18
+#: ../taxonomies/attachment-tag.php:18
 msgid "Schlagwort aktualisieren"
 msgstr "Update"
 
-#: ../assets/taxonomies/attachment-tag.php:19
+#: ../taxonomies/attachment-tag.php:19
 msgid "Neues Schlagwort erstellen"
 msgstr "Add New Tag"
 
-#: ../assets/taxonomies/attachment-tag.php:21
+#: ../taxonomies/attachment-tag.php:21
 msgid "Trenne Schlagwörter durch Kommas"
 msgstr "Separate tags with commas"
 
-#: ../assets/taxonomies/attachment-tag.php:22
+#: ../taxonomies/attachment-tag.php:22
 msgid "Hinzu"
 msgstr "Add"
 
-#: ../assets/taxonomies/attachment-tag.php:23
+#: ../taxonomies/attachment-tag.php:23
 msgid "Wähle aus den häufig genutzten Schlagwörtern"
 msgstr "Choose from the most used tags"
 
@@ -140,11 +140,11 @@ msgstr "Choose from the most used tags"
 #~ msgid "Alle Schlagworte"
 #~ msgstr "All Tags"
 
-#~ msgid "RRZE-Downloads"
-#~ msgstr "RRZE Downloads"
+#~ msgid "CMS-Basis"
+#~ msgstr "CMS Basis"
 
-#~ msgid "Einstellungen &rsaquo; RRZE-Downloads"
-#~ msgstr "Settings &rsaquo; RRZE Downloads"
+#~ msgid "Einstellungen &rsaquo; CMS-Basis"
+#~ msgstr "Settings &rsaquo; CMS Basis"
 
 #~ msgid "Feld 1"
 #~ msgstr "Field 1"
