+++ languages/rrze-downloads.pot
@@ -39,94 +39,94 @@ msgid ""
 "auf die Wordpress-Version %s."
 msgstr ""
 
-#: ..includes/Shortcode.php:219
+#: ../shortcodes/downloads.php:219
 msgid "Dowloads: Es konnten keine Dateien gefunden werden."
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:10
-#: ../assets/taxonomies/attachment-category.php:20
+#: ../taxonomies/attachment-category.php:10
+#: ../taxonomies/attachment-category.php:20
 msgid "Kategorien"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:11
+#: ../taxonomies/attachment-category.php:11
 msgid "Kategorie"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:12
+#: ../taxonomies/attachment-category.php:12
 msgid "Kategorien suchen"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:13
-#: ../assets/taxonomies/attachment-category.php:49
-#: ../assets/taxonomies/attachment-category.php:92
+#: ../taxonomies/attachment-category.php:13
+#: ../taxonomies/attachment-category.php:49
+#: ../taxonomies/attachment-category.php:92
 msgid "Alle Kategorien"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:14
+#: ../taxonomies/attachment-category.php:14
 msgid "Übergeordnet"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:15
+#: ../taxonomies/attachment-category.php:15
 msgid "Übergeordnet:"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:16
+#: ../taxonomies/attachment-category.php:16
 msgid "Kategorie bearbeiten"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:17
+#: ../taxonomies/attachment-category.php:17
 msgid "Kategorie aktualisieren"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:18
+#: ../taxonomies/attachment-category.php:18
 msgid "Neue Kategorie hinzufügen"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-category.php:19 ../assets/taxonomies/attachment-tag.php:20
+#: ../taxonomies/attachment-category.php:19 ../taxonomies/attachment-tag.php:20
 msgid "Name"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:10 ../assets/taxonomies/attachment-tag.php:24
+#: ../taxonomies/attachment-tag.php:10 ../taxonomies/attachment-tag.php:24
 msgid "Schlagwörter"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:11
+#: ../taxonomies/attachment-tag.php:11
 msgid "Schlagwort"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:12
+#: ../taxonomies/attachment-tag.php:12
 msgid "Schlagwörter suchen"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:13
+#: ../taxonomies/attachment-tag.php:13
 msgid "Beliebte Schlagwörter"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:14 ../assets/taxonomies/attachment-tag.php:53
-#: ../assets/taxonomies/attachment-tag.php:96
+#: ../taxonomies/attachment-tag.php:14 ../taxonomies/attachment-tag.php:53
+#: ../taxonomies/attachment-tag.php:96
 msgid "Alle Schlagwörter"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:17
+#: ../taxonomies/attachment-tag.php:17
 msgid "Schlagwort bearbeiten"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:18
+#: ../taxonomies/attachment-tag.php:18
 msgid "Schlagwort aktualisieren"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:19
+#: ../taxonomies/attachment-tag.php:19
 msgid "Neues Schlagwort erstellen"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:21
+#: ../taxonomies/attachment-tag.php:21
 msgid "Trenne Schlagwörter durch Kommas"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:22
+#: ../taxonomies/attachment-tag.php:22
 msgid "Hinzu"
 msgstr ""
 
-#: ../assets/taxonomies/attachment-tag.php:23
+#: ../taxonomies/attachment-tag.php:23
 msgid "Wähle aus den häufig genutzten Schlagwörtern"
 msgstr ""
