+++ taxonomies/js/media-taxonomies.js
